requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.5
certifi==2019.11.28
chardet==3.0.4
discord.py==1.3.4
dnspython==1.16.0
idna==2.9
multidict==4.7.5
Pillow==7.0.0
pycodestyle==2.5.0
pymongo==3.10.1
python-dotenv==0.12.0
requests==2.23.0
urllib3==1.25.8
websockets==8.1
yarl==1.4.2
